backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.30
pydantic>=2.5
pydantic-settings>=2.2
python-jose[cryptography]>=3.3.0
boto3>=1.34
requests>=2.32
